Rework NSMessageBuilder
This is how it's supposed to work; and it's actually useful now. You invoke it
with the target and an out-pointer to an invocation, it returns you a special
proxy that will save the invocation you make on it to that pointer.

This requires a matching change in Foundation (NSProxy).