Completely replace CFDate objects with NSDate
Apple's latest CF just dispatches all CFDate functions to NSDate, so we should do the same.

This fix is actually necessary because NSDate objects place their time interval in the same spot where CFDate objects have their CF info data. So 1) NSDate was actually not bridgable with CFDate, 2) with certain values for the time interval, calling CF functions on NSDate objects would mis-interpret the object as a CF object and cause all sorts of messes (such as mis-matched custom allocation indicators, leading to a HALT condition in CFRetain).

By the way, in case you're wondering, the value that triggered that bug was `647236405.47000003`.